backend/requirements.txt
# flayre.ai Backend Dependencies

# FastAPI & Server
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
gunicorn>=21.0.0
pydantic[email]>=2.9.2
pydantic-settings>=2.10.1

# Database
supabase>=2.10.0

# HTTP Client
httpx>=0.28.1

# Authentication
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4

# Utilities
python-multipart>=0.0.9
python-dotenv>=1.0.1
email-validator>=2.0.0